requirements.txt
Flask==3.0.0
python-docx==1.1.0
pypdf==3.17.4
docx2pdf==0.1.8
Werkzeug==3.0.1

